% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{int_times_int}
\alias{int_times_int}
\title{Multiply Input By Another Input}
\usage{
int_times_int(x, y)
}
\arguments{
\item{x}{An integer vector.}

\item{y}{An integer to multiply.}
}
\value{
An integer vector with values multiplied by \code{y}.
}
\description{
Multiply Input By Another Input
}
